revoke -m t8n --version command replace by ethereum-spec-evm --version
